src/torch_web/requirements.txt
aiosqlite==0.18.0
alembic==1.9.3
anyio==3.6.2
apiflask==1.2.3
apprise==1.2.1
asgi-lifespan==2.0.0
astroid==2.14.1
asyncpg==0.27.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.5
boto3==1.26.19
botocore==1.29.19
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.8
cryptography==39.0.1
dill==0.3.6
dnspython==2.3.0
docker==6.0.1
email-validator==1.3.1
fastapi==0.90.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security-Too==5.1.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flask-cors==3.0.10
fsspec==2023.1.0
google-auth==2.16.0
greenlet==2.0.2
griffe==0.25.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
imagehash==4.3.1
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
kubernetes==25.3.0
lazy-object-proxy==1.9.0
libsass==0.22.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
minio==7.1.12
mysql-connector-python==8.0.31
oauthlib==3.2.2
orjson==3.8.6
packaging==23.0
paramiko==2.12.0
passlib==1.7.4
pathspec==0.11.0
pbkdf2==1.3
pendulum==2.1.2
Pillow==9.3.0
pip==23.0
platformdirs==3.0.0
prefect==2.6.9
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pylint==2.16.1
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-slugify==8.0.0
python-socketio==5.7.2
pytz==2022.7.1
pytzdata==2020.1
pywin32==305; sys_platform == 'win32'
PyYAML==6.0
pyzbar==0.1.9
readchar==4.0.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
setuptools==58.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.23.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
websocket-client==1.5.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.12.1
